Defer all format-on-save operations to null-ls
Prevent any language servers from formatting files and defer this
responsibility to null-ls. Reasons:

  1. Formatting is explicitly handled in one place (sometimes it's not
     obvious which language servers come bundled with formatting). If
     formatting breaks in some way, I know exactly where to look.
  2. Language servers won't always be available in certain situations,
     projects, filetypes, etc. There might be a case where I want
     formatting, but don't want to run an entire language server for it.

If a particular language server is the only option, we can always enable
it specifically for that language via `language_server.setup({})`

Also added a StyLua config file for this repo. The next few nvim
commits are going to have formatting changes bundled in :)