Fix format-on-save
Format-on-save was still behaving inconsistently, sometimes it would
format on write, other times it wouldn't and would require a manual call
to `vim.lsp.buf.format()`

Now we're using lsp-format to handle format-on-save, which attempts to
solve the complications that arise when dealing with asynchronous
formatting, writing to buffers and multiple formatters attached to a
single buffer. This plugin should deal with all of that while making the
setup easer.

Refs:
https://github.com/lukas-reineke/lsp-format.nvim
https://www.reddit.com/r/neovim/comments/t4tgpr/formatnvim_becomes_lspformatnvim/